backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
numpy
realesrgan
requests
aiofiles
